Alter Compile function to work with M2 >=2.1
The release of Magento 2.1.0 removed setup:di:compile-multi-tenant. As we haven't initiated Magento in this script we can't check its version number. To get round this I've put in some error handling. First we try multi-tenant complilation and if that fails we catch as we are greater or equal to 2.1 and use setup:di:compile.